Consolidate transaction endpoints to single unified POST handler, migrate reward points to BigDecimal, and enhance controller test parameter assertion matching

+++ src/main/java/com/retailer/rewards/dto/MonthlyRewardsSummary.java
@@ -5,6 +5,7 @@ import lombok.Builder;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
+import java.math.BigDecimal;
 import java.util.List;
 
 /**
@@ -26,7 +27,7 @@ public class MonthlyRewardsSummary {
     private String monthName;
 
     /** Total reward points earned this month. */
-    private long monthlyPoints;
+    private BigDecimal monthlyPoints;
 
     /** Breakdown per transaction. */
     private List<TransactionRewardDetail> transactions;
